+++ src/interfaces/tooltipNew.d.ts
@@ -3,11 +3,14 @@ import * as PopperJS from "popper.js";
 
 type RefHandler = (ref: HTMLElement | null) => void;
 
+export type TooltipTheme = "normal" | "invert" | "important";
+
 export type TooltipTriggerAction = "managed" | "click" | "hover";
 
 export interface IPopperTooltipProps
   extends React.HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
+  theme?: TooltipTheme;
   closeOnOutsideClick?: boolean;
   eventsEnabled?: boolean;
   hoverOutDelayTimeout?: number;
@@ -27,6 +30,7 @@ export interface IPopperTooltipProps
     | React.ReactNode;
   triggerActionType?: TooltipTriggerAction;
   zIndex: number;
+  onOpen?: () => void;
 }
 
 type CssTooltipPlacement =
@@ -46,6 +50,7 @@ type CssTooltipPlacement =
 export interface ICssTooltipProps extends React.HTMLAttributes<HTMLDivElement> {
   arrowClassName?: string;
   children: React.ReactNode;
+  theme?: TooltipTheme;
   isVisible?: boolean;
   arrowOffsetTop?: string;
   arrowOffsetBottom?: string;
@@ -60,5 +65,19 @@ export interface ICssTooltipProps extends React.HTMLAttributes<HTMLDivElement> {
   zIndex?: number;
 }
 
+export interface IUserGuideTooltipProps
+  extends React.HTMLAttributes<HTMLDivElement> {
+  children: React.ReactNode;
+  zIndex: number;
+  element: React.ReactNode | Element;
+  scrollableWrapper?: Element;
+  isVisible?: boolean;
+  slide?: boolean;
+  theme?: TooltipTheme;
+  placement?: PopperJS.Placement;
+  containerName?: string;
+}
+
 export var PopperTooltip: React.ComponentType<IPopperTooltipProps>;
 export var CssTooltip: React.ComponentType<ICssTooltipProps>;
+export var UserGuideTooltip: React.ComponentType<IUserGuideTooltipProps>;