+++ styleguide.config.js
@@ -38,7 +38,10 @@ module.exports = {
         },
         {
           name: 'Button',
-          components: './src/components/Button/Button.js'
+          components: [
+            './src/components/Button/Button.js',
+            './src/components/ButtonGroup/ButtonGroup.js'
+          ]
         },
         {
           name: 'Tab',
@@ -70,7 +73,8 @@ module.exports = {
           content: './src/components/TooltipNew/design-guide.md',
           components: [
             './src/components/TooltipNew/PopperTooltip.js',
-            './src/components/TooltipNew/CssTooltip.js'
+            './src/components/TooltipNew/CssTooltip.js',
+            './src/components/TooltipNew/UserGuideTooltip.js'
           ]
         },
         {
@@ -243,7 +247,7 @@ module.exports = {
           .view-html-btn--active {
             border-bottom: 2px #f28a25 solid;
           }
-          
+
           pre.prettyprint { display: block; background-color: #0a001f; font-size: 13px; font-family: monospace; margin: 0; padding: 8px 16px;}
           pre .tag { color: #99b2ff; }
           pre .nocode { background-color: transparent; color: #000 }